Refactor to flat structure: split _core.py into hmm.py and algorithms.py
