+++ src/hmm/__init__.py
@@ -1,6 +1,7 @@
 # HMMPY - Hidden Markov Models in Python
 # Based on Rabiner tutorial
 
-from hmm._core import HMM, HMMClassifier, backward, baum_welch, forward, viterbi
+from hmm.algorithms import backward, baum_welch, forward, viterbi
+from hmm.hmm import HMM, HMMClassifier
 
 __all__ = ["HMM", "HMMClassifier", "forward", "backward", "viterbi", "baum_welch"]
